Gestion des vitesses différentes pour chaque switch Distinction ES/Switch Passage de la gestion du temps au 1/10 e de ms

+++ Artemis_Core/src/root/elements/network/modules/task/ISchedulable.java
@@ -34,10 +34,10 @@ public interface ISchedulable {
 	void setCurrentPeriod(int period);
 	
 	/* Next send getter */
-	int getNextSend();
+	double getNextSend();
 	
 	/* Next send setter */
-	void setNextSend(int nextSend);
+	void setNextSend(double nextSend);
 	
 	/* Network path getter */
 	Vector<NetworkAddress> getNetworkPath();
